Fiber

+++ internal/service/chat.service.go
@@ -17,7 +17,7 @@ func NewChatService(repository repository.Chat) *ChatService {
 	}
 }
 
-func (service *ChatService) CreateChat(ctx context.Context, userId int64, chatDTO Entities.ChatDTO) (int64, error) {
+func (service *ChatService) CreateChat(ctx context.Context, userId uint, chatDTO Entities.ChatDTO) (uint, error) {
 	if len(chatDTO.Members) < 2 {
 		return 0, errors.New("members must be more than 1")
 	}
@@ -35,18 +35,18 @@ func (service *ChatService) CreateChat(ctx context.Context, userId int64, chatDT
 	return service.repository.CreateChat(ctx, chatDTO)
 }
 
-func (service *ChatService) UpdateChat(ctx context.Context, userId int64, chatId int64, chatDTO Entities.ChatUpdateDTO) error {
+func (service *ChatService) UpdateChat(ctx context.Context, userId uint, chatId uint, chatDTO Entities.ChatUpdateDTO) error {
 	return service.repository.UpdateChat(ctx, userId, chatId, chatDTO)
 }
 
-func (service *ChatService) DeleteChat(ctx context.Context, userId int64, chatId int64) ([]int64, error) {
+func (service *ChatService) DeleteChat(ctx context.Context, userId uint, chatId uint) ([]uint, error) {
 	return service.repository.DeleteChat(ctx, userId, chatId)
 }
 
-func (service *ChatService) GetChats(ctx context.Context, userId int64) (string, string, string, []int64, []int64, string, []Entities.Chat, error) {
+func (service *ChatService) GetChats(ctx context.Context, userId uint) (string, string, string, []uint, []uint, string, []Entities.Chat, error) {
 	return service.repository.GetChats(ctx, userId)
 }
 
-func (service *ChatService) UpdateChatLists(ctx context.Context, id int64, newChatLists string) error {
+func (service *ChatService) UpdateChatLists(ctx context.Context, id uint, newChatLists string) error {
 	return service.repository.UpdateChatLists(ctx, id, newChatLists)
 }
